refactor saving words in library. Catch scenario when library can not contain two words with similar spelling
