+++ src/main/kotlin/systems/ajax/englishstudytelegrambot/exception/WordExceptions.kt
@@ -2,4 +2,4 @@ package systems.ajax.englishstudytelegrambot.exception
 
 class WordNotFoundBySpendingException : RuntimeException()
 
-class WordAlreadyPresentException : RuntimeException()
+class WordAlreadyPresentInLibraryException : RuntimeException()
