+++ src/main/kotlin/systems/ajax/englishstudytelegrambot/repository/WordRepository.kt
@@ -1,18 +1,18 @@
 package systems.ajax.englishstudytelegrambot.repository
 
+import org.springframework.data.mongodb.core.FindAndModifyOptions
 import org.springframework.data.mongodb.core.MongoTemplate
+import org.springframework.data.mongodb.core.query.Criteria
+import org.springframework.data.mongodb.core.query.Query
 import org.springframework.data.mongodb.core.query.Update
 import org.springframework.stereotype.Repository
 import systems.ajax.englishstudytelegrambot.entity.Library
 import systems.ajax.englishstudytelegrambot.entity.Word
 import systems.ajax.englishstudytelegrambot.exception.LibraryIsMissingException
-import systems.ajax.englishstudytelegrambot.exception.WordNotFoundBySpendingException
 
 interface WordRepository {
 
-    fun saveNewWordInLibrary(word: Word, libraryName: String, telegramUserId: String): Library
-
-    fun findById(wordSpelling: String): Word
+    fun saveNewWordInLibrary(word: Word, library: Library): Library
 }
 
 @Repository
@@ -20,13 +20,11 @@ class WordRepositoryImpl(
     val mongoTemplate: MongoTemplate
 ) : WordRepository {
 
-    override fun saveNewWordInLibrary(word: Word, libraryName: String, telegramUserId: String): Library =
+    override fun saveNewWordInLibrary(word: Word, library: Library): Library =
         mongoTemplate.findAndModify(
-            CommonQuery.queryToFindLibraryByNameAndTelegramUserId(telegramUserId, libraryName),
+            Query.query(Criteria.where("_id").`is`(library.id)),
             Update().push("words", word),
+            FindAndModifyOptions().returnNew(true),
             Library::class.java
-        )?: throw LibraryIsMissingException()
-
-    override fun findById(wordSpelling: String): Word =
-        mongoTemplate.findById(wordSpelling, Word::class.java)?: throw WordNotFoundBySpendingException()
+        ) ?: throw LibraryIsMissingException()
 }
