+++ src/main/kotlin/systems/ajax/englishstudytelegrambot/service/WordService.kt
@@ -5,12 +5,11 @@ import org.springframework.stereotype.Service
 import systems.ajax.englishstudytelegrambot.dto.WordDto
 import systems.ajax.englishstudytelegrambot.entity.Library
 import systems.ajax.englishstudytelegrambot.entity.Word
+import systems.ajax.englishstudytelegrambot.exception.WordAlreadyPresentInLibraryException
 import systems.ajax.englishstudytelegrambot.repository.LibraryRepository
 import systems.ajax.englishstudytelegrambot.repository.WordRepository
 
 interface WordService {
-
-    fun getWordByItsSpelling(wordSpelling: String): Word
 
     suspend fun saveNewWordInLibrary(libraryName: String, telegramUserId: String, wordDto: WordDto): Library
 }
@@ -18,21 +17,30 @@ interface WordService {
 @Service
 class WordServiceImpl(
     val wordRepository: WordRepository,
+    val libraryRepository: LibraryRepository,
     val additionalInfoAboutWordService: AdditionalInfoAboutWordService
 ) : WordService {
 
     override suspend fun saveNewWordInLibrary(libraryName: String, telegramUserId: String, wordDto: WordDto): Library {
-        val word =
-            Word(
-                wordDto.spelling,
-                wordDto.translate,
-                additionalInfoAboutWordService.findAdditionInfoAboutWord(wordDto.spelling)
-            )
-        return wordRepository.saveNewWordInLibrary(word, libraryName, telegramUserId)
+        val library = libraryRepository.getLibraryByPairLibraryNameAndTelegramUserId(libraryName, telegramUserId)
+        library
+            .takeIf { isLibraryContainsWordWithSpelling(it, wordDto.spelling) }
+            ?.let {
+                val word =
+                    Word(
+                        wordDto.spelling,
+                        wordDto.translate,
+                        additionalInfoAboutWordService.findAdditionInfoAboutWord(wordDto.spelling)
+                    )
+                return wordRepository.saveNewWordInLibrary(word, it)
+            } ?: throw WordAlreadyPresentInLibraryException()
     }
 
-    override fun getWordByItsSpelling(wordSpelling: String): Word =
-        wordRepository.findById(wordSpelling)
+
+    private fun isLibraryContainsWordWithSpelling(
+        library: Library,
+        wordSpelling: String
+    ) = library.words.none { word -> word.spelling == wordSpelling }
 
     companion object {
         val log = LoggerFactory.getLogger(this::class.java)
